feat!: return raw search results
Before, we were returning a collection of found pages, with a `hit()` helper to retrieve the raw loupe result by the page's uuid. But this approach doesn't work for multiple hits with the same page. So for now we just return the loupe hits as a collection. Maybe we can think of an abstraction that feels more kirby-like in the future.